clean up code so it is not as messy
